move broken test for inheritance to jvm source set
